Move GetSyncCallback to the Commons package; Name ChannelListener thread
